Sprint No.16 AudioPlayer to MVVM p2 fix bag current TrackTime 0.30

+++ app/src/main/java/com/example/playlistmaker/player/domain/usecase/GetAudioPositionUseCase.kt
@@ -3,5 +3,5 @@ package com.example.playlistmaker.player.domain.usecase
 import com.example.playlistmaker.player.domain.repository.AudioPlayerRepository
 
 class GetAudioPositionUseCase(private val repository: AudioPlayerRepository) {
-    fun execute(): Int = repository.getCurrentPosition()
+    fun execute(): Long = repository.getCurrentPosition()  // было Int
 }
